Recalculate bounds without the needless request layout calls
